Fix: mypage > userPage foldername renamed
