--- constants/mypage.tsx
@@ -1,25 +1,0 @@
-export const mypage_menus = [
-  { id: 1, name: "계정관리", path: "/myPage/myAccount" },
-  { id: 2, name: "FAQ", path: "/myPage/faq" },
-  { id: 3, name: "서비스 피드백", path: "/myPage/feedback" },
-  { id: 4, name: "약관 및 정책", path: "/myPage/policy" },
-];
-
-export const myaccount_menus = [
-  { id: 1, name: "휴대폰 번호 변경", path: "/myPage/myAccount/phone" },
-  { id: 2, name: "비밀번호 변경", path: "/myPage/myAccount/pswd" },
-  { id: 3, name: "로그아웃", path: "" },
-];
-
-export const policy_menus = [
-  {
-    id: 1,
-    name: "서비스 이용 약관",
-    path: "/myPage/policy/term",
-  },
-  {
-    id: 2,
-    name: "개인정보 처리 방침",
-    path: "/myPage/policy/privacy",
-  },
-];
