+++ constants/userPage.tsx
@@ -1,0 +1,25 @@
+export const userPage_menus = [
+  { id: 1, name: "계정관리", path: "/userPage/myAccount" },
+  { id: 2, name: "FAQ", path: "/userPage/faq" },
+  { id: 3, name: "서비스 피드백", path: "/userPage/feedback" },
+  { id: 4, name: "약관 및 정책", path: "/userPage/policy" },
+];
+
+export const myaccount_menus = [
+  { id: 1, name: "휴대폰 번호 변경", path: "/userPage/myAccount/phone" },
+  { id: 2, name: "비밀번호 변경", path: "/userPage/myAccount/pswd" },
+  { id: 3, name: "로그아웃", path: "" },
+];
+
+export const policy_menus = [
+  {
+    id: 1,
+    name: "서비스 이용 약관",
+    path: "/userPage/policy/term",
+  },
+  {
+    id: 2,
+    name: "개인정보 처리 방침",
+    path: "/userPage/policy/privacy",
+  },
+];
