--- pages/mypage/feedback/index.js
@@ -1,13 +1,0 @@
-const Feedback = () => {
-  return (
-    <div>
-      <iframe
-        src="https://7xrdp4cp24a.typeform.com/to/Lkh7C9zd"
-        width="100%"
-        height="500px"
-      ></iframe>
-    </div>
-  );
-};
-
-export default Feedback;
