+++ pages/userPage/policy/privacy/index.js
