+++ pages/userPage/policy/term/index.js
